fix(firmware): safe tech debt fixes C-1/C-2/C-3/H-2/H-3
No runtime behaviour change. Type safety and debug log removal only.

- C-1: StateStore::getCurrentEffect() return uint8_t → EffectId (uint16_t)
- C-2: WsEffectsCommands dangling pointer — familyNameBuf moved outside loop
- C-3: V1ApiRoutes EffectId truncation — uint8_t → uint16_t in 5 endpoints
- H-2: Remove blocking Serial.printf/println from TransitionEngine render path
- H-3: Remove agent-log Serial.printf from LGPSpectrumDetailEffect render path
- H-3: Remove g_bloomDebugEnabled debug machinery from BeatPulseBloomEffect

+++ firmware-v3/src/network/webserver/V1ApiRoutes.cpp
@@ -402,7 +402,7 @@ void V1ApiRoutes::registerRoutes(
                               ErrorCodes::MISSING_FIELD, "Missing id parameter", "id");
             return;
         }
-        uint8_t id = request->getParam("id")->value().toInt();
+        uint16_t id = request->getParam("id")->value().toInt();
         handlers::AudioHandlers::handleMappingsGet(request, id, ctx.renderer);
     });
 
@@ -417,7 +417,7 @@ void V1ApiRoutes::registerRoutes(
                                   ErrorCodes::MISSING_FIELD, "Missing id parameter", "id");
                 return;
             }
-            uint8_t id = request->getParam("id")->value().toInt();
+            uint16_t id = request->getParam("id")->value().toInt();
             handlers::AudioHandlers::handleMappingsSet(request, id, data, len, ctx.renderer);
         }
     );
@@ -430,7 +430,7 @@ void V1ApiRoutes::registerRoutes(
                               ErrorCodes::MISSING_FIELD, "Missing id parameter", "id");
             return;
         }
-        uint8_t id = request->getParam("id")->value().toInt();
+        uint16_t id = request->getParam("id")->value().toInt();
         handlers::AudioHandlers::handleMappingsDelete(request, id);
     });
 
@@ -442,7 +442,7 @@ void V1ApiRoutes::registerRoutes(
                               ErrorCodes::MISSING_FIELD, "Missing id parameter", "id");
             return;
         }
-        uint8_t id = request->getParam("id")->value().toInt();
+        uint16_t id = request->getParam("id")->value().toInt();
         handlers::AudioHandlers::handleMappingsEnable(request, id, true);
     });
 
@@ -454,7 +454,7 @@ void V1ApiRoutes::registerRoutes(
                               ErrorCodes::MISSING_FIELD, "Missing id parameter", "id");
             return;
         }
-        uint8_t id = request->getParam("id")->value().toInt();
+        uint16_t id = request->getParam("id")->value().toInt();
         handlers::AudioHandlers::handleMappingsEnable(request, id, false);
     });
